docker/test/requirements.txt
testcontainers==3.7.1

sqlmodel==0.0.8
psycopg2-binary==2.9.5

requests==2.28.2

pytest==7.2.2
